# Conflicts:
#	Test/CombinedTest/Core/BeltConveyorTest.cs
#	Test/CombinedTest/Core/BlockConfigTest.cs
#	Test/CombinedTest/Core/ChestLogicTest.cs
#	Test/CombinedTest/Core/Generate/MachineIOGenerate.cs
#	Test/CombinedTest/Core/MachineIOTest.cs
#	Test/CombinedTest/Core/MinerCanBeMinedTest.cs
#	Test/CombinedTest/Core/MinerMiningTest.cs
#	Test/CombinedTest/Core/PowerGeneratorTest.cs
#	Test/CombinedTest/Game/ConnectElectricSegmentTest.cs
#	Test/CombinedTest/Game/DisconnectElectricSegmentTest.cs
#	Test/CombinedTest/Game/InventoryItemMoveServiceTest.cs
#	Test/CombinedTest/Game/SaveJsonFileTest.cs
#	Test/CombinedTest/Game/VeinGeneratorTest.cs
#	Test/CombinedTest/Server/PacketTest/CraftProtocolFailedTest.cs
#	Test/CombinedTest/Server/PacketTest/CraftProtocolTest.cs
#	Test/CombinedTest/Server/PacketTest/Event/BlockInventoryUpdateEventPacketTest.cs
#	Test/CombinedTest/Server/PacketTest/Event/BlockPlaceEventPacketTest.cs
#	Test/CombinedTest/Server/PacketTest/Event/BlockRemoveEventPacketTest.cs
#	Test/CombinedTest/Server/PacketTest/Event/CraftInventoryUpdateTest.cs
#	Test/CombinedTest/Server/PacketTest/Event/PlayerMainInventoryUpdateTest.cs
#	Test/CombinedTest/Server/PacketTest/InitialHandshakeProtocolTest.cs
#	Test/CombinedTest/Server/PacketTest/InventoryItemMoveProtocolTest.cs
#	Test/CombinedTest/Server/PacketTest/MiningOperationProtocolTest.cs
#	Test/CombinedTest/Server/PacketTest/PlaceHotBarBlockProtocolTest.cs
#	Test/CombinedTest/Server/PacketTest/PlayerCoordinateSendProtocolTest.cs
#	Test/CombinedTest/Server/PacketTest/PlayerInventoryProtocolTest.cs
#	Test/CombinedTest/Server/PacketTest/PutBlockProtocolTest.cs
#	Test/CombinedTest/Server/PacketTest/RemoveBlockProtocolTest.cs
#	Test/CombinedTest/Server/PacketTest/RequestBlockInventoryTest.cs
#	Test/CombinedTest/Server/PacketTest/SendCommandProtocolTest.cs
#	Test/UnitTest/Core/Block/MachineRecipeConfigTest.cs
#	Test/UnitTest/Core/Block/MinerSaveLoadTest.cs
#	Test/UnitTest/Core/Other/ConnectingInventoryListPriorityInsertItemServiceTest.cs
#	Test/UnitTest/Core/Other/ItemConfigTest.cs
#	Test/UnitTest/Core/Other/ItemStackTest.cs
#	Test/UnitTest/Core/Other/NullRecipeDataTest.cs
#	Test/UnitTest/Core/Other/OreConfigTest.cs
#	Test/UnitTest/Game/BlockPlaceToConnectionBlockTest.cs
#	Test/UnitTest/Game/CraftingInventoryDataTest.cs
#	Test/UnitTest/Game/SaveLoad/AssembleEntitySaveJsonTextTest.cs
#	Test/UnitTest/Game/SaveLoad/AssemblePlayerInventorySaveJsonTextTest.cs
#	Test/UnitTest/Game/SaveLoad/AssembleSaveJsonTextTest.cs
#	Test/UnitTest/Game/SaveLoad/BeltConveyorSaveLoadTest.cs
#	Test/UnitTest/Game/SaveLoad/ChestSaveLoadTest.cs
#	Test/UnitTest/Game/SaveLoad/MachineSaveLoadTest.cs
#	Test/UnitTest/Game/SaveLoad/PowerGeneratorSaveLoadTest.cs
#	Test/UnitTest/Game/WorldInstallationDatastoreTest.cs
#	Test/UnitTest/Mod/ModGetConfigStringTest.cs
#	Test/UnitTest/Server/Player/CoordinateToChunkBlocksTest.cs

+++ Core.Block/Config/LoadConfig/BlockConfigJsonLoad.cs
@@ -88,19 +88,19 @@ namespace Core.Block.Config.LoadConfig
                 if (modelTransformJson.pos != null)
                 {
                     var pos = modelTransformJson.pos;
-                    posVector3 = new CoreVector3(pos[0], pos[1], pos[2]);
+                    posVector3 = new CoreVector3((float)pos[0], (float)pos[1],(float) pos[2]);
                 }
                 
                 
                 if (modelTransformJson.rot != null)
                 {
                     var rot = modelTransformJson.rot;
-                    rotVector3 = new CoreVector3(rot[0], rot[1], rot[2]);
+                    rotVector3 = new CoreVector3((float)rot[0],(float) rot[1], (float)rot[2]);
                 }
                 if (modelTransformJson.scale != null)
                 {
                     var scale = modelTransformJson.scale;
-                    scaleVector3 = new CoreVector3(scale[0], scale[1], scale[2]);
+                    scaleVector3 = new CoreVector3((float)scale[0], (float)scale[1], (float)scale[2]);
                 }
                 
                 modelTransform = new ModelTransform(posVector3, rotVector3, scaleVector3);
